msm_vidc: update driver for firmware interface changes
This change updates the driver to the new firmware requirements.
Since, FW interface has changed significantly so pre-existing driver code
does not work with this firmware. This change is a must for driver to
work with new firmware. After this change, old FW will not work with this
driver.

Change-Id: I2f8408dd46909f7f527252410655c6fa919c465a
Signed-off-by: Ashray Kulkarni <ashrayk@codeaurora.org>
Signed-off-by: Vinay Kalia <vkalia@codeaurora.org>

Conflicts:
	drivers/media/platform/msm/vidc/msm_vidc_common.c
	drivers/media/platform/msm/vidc/vidc_hal.c
	drivers/media/platform/msm/vidc/vidc_hal.h
	drivers/media/platform/msm/vidc/vidc_hal_helper.h
	drivers/media/platform/msm/vidc/vidc_hal_io.h
	include/media/msm_media_info.h