msm: camera: add support for bus overflow recovery
When bus overflow is detected, system can enter bad state. So we
halt, reset and restart the AXI Bus Bridge, CAMIF and ISPIF INTF
in order to recover.
On overflow detection, kernel halts the AXI Bus Bridge and sends ERROR
message to user spacce and wait for it to trigger reset and restart.

Change-Id: Ie473d230d4f64b6fd3a0d98f181f4c82a48dab03
Signed-off-by: Peter Liu <pingchie@codeaurora.org>

Conflicts:
	drivers/media/platform/msm/camera_v2/isp/msm_isp.h
	drivers/media/platform/msm/camera_v2/isp/msm_isp40.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp44.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp46.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp_axi_util.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp_axi_util.h
	drivers/media/platform/msm/camera_v2/isp/msm_isp_util.c
	drivers/media/platform/msm/camera_v2/ispif/msm_ispif.c
	include/media/msmb_isp.h