qcom: sysmon: Replace sending strings as parameters to sysmon APIs
Currently, all the sysmon APIs use string names to derive the
information about subsystems that the events should be sent to.
To do so, a list of subsystems that support sysmon are stored
in a static array. This approach makes sysmon break when the
name of a subsystem changes or when new subsystems are added.

Instead, store the sysmon pdev id information that is read
from the subsystem's dt nodes in the subsys descriptor. These
descriptors are then passed as parameters to the sysmon APIs.

Change-Id: I0c0b68412acb4913e823586931a7fd0b8f03f15b
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>

Conflicts:
	arch/arm/boot/dts/qcom/msm8994.dtsi
	drivers/esoc/esoc-mdm-4x.c
	drivers/soc/qcom/subsystem_restart.c
	drivers/soc/qcom/sysmon.c
	include/soc/qcom/sysmon.h