qcom: ssr: Reduce sysmon event response timeout
Currently, we wait for 5 sec to allow the subsystem sysmon
server to respond after sending a sysmon event. This timeout
can be reduced to a much lesser value to prevent additional
wait time during subsystem bootup. Set the new timeout limit
to 500 msec which should be sufficient.
Also, avoid sending sysmon events to subsystems which are not
online.

Change-Id: I3ffb81d7e8d38370e3504c81fec9c6a40fdab30c
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>

Conflicts:
	drivers/soc/qcom/subsystem_restart.c